Revert "refactor: use @tanstack/solid-query for user profile API calls"
This reverts commit e471ac745cb6fa3c99e49c4e5b09b5f15663c023.
