+++ client/src/core/context/user/provider.tsx
@@ -1,5 +1,4 @@
 import { AuthnContext } from "core/context/authn";
-import { UiContext } from "core/context/ui";
 import { createUserContext } from "core/context/initializers";
 import { useContext } from "core/context/utils";
 import {
@@ -17,7 +16,6 @@ export const UserContext = createContext<Accessor<UserSvc>>(createUserContext);
 export const UserProvider: Component<ParentProps> = props => {
 	const value = useStore();
 	const authnContext = useContext(AuthnContext);
-	const uiContext = useContext(UiContext);
 
 	// TODO: once BE is up properly this should not be dependent on `authn`
 	// `authn` just authenticates
@@ -27,9 +25,8 @@ export const UserProvider: Component<ParentProps> = props => {
 		}
 
 		const profile = authnContext().profile;
-		const queryClient = uiContext().queryClient;
 
-		value().actions.init(authnContext().keycloak?.token, profile, queryClient);
+		value().actions.init(authnContext().keycloak?.token, profile);
 	});
 
 	return (
