+++ client/src/feat/profile/hooks/use-profile-form.ts
@@ -1,6 +1,5 @@
 import { createForm, reset, type SubmitHandler } from "@modular-forms/solid";
-import { useMutation, useQueryClient } from "@tanstack/solid-query";
-import { queryKeys } from "core/constants/query-keys";
+import { useMutation } from "@tanstack/solid-query";
 import { AuthnContext } from "core/context/authn";
 import { UserContext } from "core/context/user";
 import { useContext } from "core/context/utils";
@@ -14,7 +13,6 @@ export const useProfileForm = () => {
 	const profileContext = useContext(ProfileContext);
 	const userContext = useContext(UserContext);
 	const authnContext = useContext(AuthnContext);
-	const queryClient = useQueryClient();
 
 	const [form, { Form, Field, FieldArray }] = createForm<ProfileForm>({
 		validateOn: "input",
@@ -29,18 +27,10 @@ export const useProfileForm = () => {
 		onSuccess: async () => {
 			const token = authnContext().keycloak?.token;
 			const profile = authnContext().profile;
-			const userProfile = userContext().profile;
 			
-			// Invalidate user profile queries to trigger refetch
+			// Refresh user profile data after successful update
 			if (token && profile?.email) {
-				await queryClient.invalidateQueries({
-					queryKey: queryKeys.getUserProfile(token, profile.email),
-				});
-			}
-			if (token && userProfile?.uuid) {
-				await queryClient.invalidateQueries({
-					queryKey: queryKeys.getUserPersonalDetails(token, userProfile.uuid),
-				});
+				await userContext().actions.refreshProfile(token, profile.email);
 			}
 		},
 	}));
